Migrate quote unquote logic to evaluator expand file
