--- src/evaluator/quote.rs
@@ -1,3 +1,0 @@
-pub fn quote() {
-
-}
